Remove duplicate code and rationalise pianoteq engine
