+++ zyngine/__init__.py
@@ -12,6 +12,7 @@ __all__ = [
 	"zynthian_engine_setbfree",
 	"zynthian_engine_modui",
 	"zynthian_engine_pianoteq",
+	"zynthian_engine_pianoteq6",
 	"zynthian_engine_puredata",
 	"zynthian_engine_aeolus",
 	"zynthian_engine_jalv",
@@ -35,6 +36,7 @@ from zyngine.zynthian_engine_fluidsynth import *
 from zyngine.zynthian_engine_setbfree import *
 from zyngine.zynthian_engine_modui import *
 from zyngine.zynthian_engine_pianoteq import *
+from zyngine.zynthian_engine_pianoteq6 import *
 from zyngine.zynthian_engine_puredata import *
 from zyngine.zynthian_engine_aeolus import *
 from zyngine.zynthian_engine_jalv import *
