+++ zyngui/zynthian_gui_engine.py
@@ -31,7 +31,7 @@ from collections import OrderedDict
 # Zynthian specific modules
 from zyngine import *
 from zyngine.zynthian_engine_pianoteq import *
-from zyngine.zynthian_engine_pianoteq_jsonrpc import *
+from zyngine.zynthian_engine_pianoteq6 import *
 from zyngine.zynthian_engine_jalv import *
 from zyngine.zynthian_engine_sooperlooper import zynthian_engine_sooperlooper
 from zyngui import zynthian_gui_config
@@ -70,16 +70,16 @@ class zynthian_gui_engine(zynthian_gui_selector):
 			['MD', ("MOD-UI", "MOD-UI - Plugin Host", "Special", None, zynthian_engine_modui, True)]
 		])
 
-		if check_pianoteq_binary():
-			pianoteq_title="Pianoteq {}.{} {}{}".format(
-				PIANOTEQ_VERSION[0],
-				PIANOTEQ_VERSION[1],
-				PIANOTEQ_PRODUCT,
-				" (Demo)" if PIANOTEQ_TRIAL else "")
-			if PIANOTEQ_VERSION[0] >= 7 and PIANOTEQ_VERSION[1] >= 5:
-				cls.engine_info['PT'] = (PIANOTEQ_NAME, pianoteq_title, "MIDI Synth", None, zynthian_engine_pianoteq_jsonrpc, True)
+		pt_info = get_pianoteq_binary_info()
+		if pt_info:
+			pianoteq_title = "Pianoteq {} {}{}".format(
+				pt_info['version_str'],
+				pt_info['product'],
+				" (Demo)" if pt_info['trial'] else "")
+			if pt_info['api']:
+				cls.engine_info['PT'] = ('Pianoteq', pianoteq_title, "MIDI Synth", None, zynthian_engine_pianoteq, True)
 			else:
-				cls.engine_info['PT'] = (PIANOTEQ_NAME, pianoteq_title, "MIDI Synth", None, zynthian_engine_pianoteq, True)
+				cls.engine_info['PT'] = ('Pianoteq', pianoteq_title, "MIDI Synth", None, zynthian_engine_pianoteq6, True)
 		
 		for plugin_name, plugin_info in get_jalv_plugins().items():
 			eng = 'JV/{}'.format(plugin_name)
